custom_wide: make menu button visible on followed profiles
The chaos-social-wide theme has less horizontal space in the columns, so the menu button on followed profiles is pushed off the screen by the larger "Unfollow" button and the bell button beside it. Remove the min-width constraint on the Unfollow button, whose text will be broken onto to lines, to fix this.